scripts/requirements.txt
supabase==2.10.0
anthropic==0.40.0
google-api-python-client==2.155.0
google-auth==2.36.0
python-dotenv==1.0.1
pandas==2.2.0
openpyxl==3.1.2
requests==2.31.0
